Migrate PriceTicker to shadcn registry (#22)
- Update PriceTicker to use components from shadcn registry
- Add components.json for shadcn configuration
- Add src/lib/utils.ts for cn utility
- Update installation instructions
