+++ src/app/docs/components/price-ticker/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useState, useCallback } from "react";
-import { PriceTicker } from "./component";
+import { PriceTicker } from "@/components/w3-kit/price-ticker";
 import { Code, Eye } from "lucide-react";
 import { CodeBlock } from "@/components/docs/codeBlock";
 
@@ -233,7 +233,7 @@ export default function Page() {
                   <p className="text-sm text-gray-600 dark:text-gray-400 mb-4">
                     Run the following command to add the Price Ticker component to your project:
                   </p>
-                  <CodeBlock code="npx w3-kit@latest add price-ticker" id="cli" />
+                  <CodeBlock code="npx shadcn@latest add https://w3-kit.com/registry/price-ticker.json" id="cli" />
                   <p className="text-sm text-gray-600 dark:text-gray-400 mt-4">
                     This will:
                   </p>
